+++ ios/Pods/FBSDKCoreKit/FBSDKCoreKit/FBSDKCoreKit/Internal/BridgeAPI/FBSDKBridgeAPICrypto.h
@@ -22,6 +22,9 @@
 
 @interface FBSDKBridgeAPICrypto : NSObject
 
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
+
 + (void)addCipherKeyToQueryParameters:(NSMutableDictionary *)queryParameters;
 + (NSDictionary *)decryptResponseForRequest:(FBSDKBridgeAPIRequest *)request
                             queryParameters:(NSDictionary *)queryParameters