+++ ios/Pods/FBSDKCoreKit/FBSDKCoreKit/FBSDKCoreKit/Internal/BridgeAPI/FBSDKBridgeAPIResponse.m
@@ -23,7 +23,6 @@
 #import "FBSDKBridgeAPIProtocolType.h"
 #import "FBSDKBridgeAPIRequest+Private.h"
 #import "FBSDKInternalUtility.h"
-#import "FBSDKMacros.h"
 #import "FBSDKTypeUtility.h"
 #import "FBSDKUtility.h"
 
@@ -117,12 +116,6 @@ NS_DESIGNATED_INITIALIZER;
     _error = [error copy];
   }
   return self;
-}
-
-- (instancetype)init
-{
-  FBSDK_NOT_DESIGNATED_INITIALIZER(initWithRequest:responseParameters:cancelled:error:);
-  return [self initWithRequest:nil responseParameters:nil cancelled:NO error:nil];
 }
 
 #pragma mark - NSCopying