+++ ios/Pods/FBSDKShareKit/FBSDKShareKit/FBSDKShareKit/Internal/FBSDKLikeActionControllerCache.m
@@ -45,12 +45,6 @@
   return self;
 }
 
-- (instancetype)init
-{
-  FBSDK_NOT_DESIGNATED_INITIALIZER(initWithAccessTokenString:);
-  return [self initWithAccessTokenString:nil];
-}
-
 #pragma mark - NSCoding
 
 + (BOOL)supportsSecureCoding